Change some methods to allow the execution of the model
Methods changed in the Gatherer class: directory_content,
convert_nfcapd_csv, open_csv and clean_files.
Methods changed in the Formatter class: format_flows and
change_data_types.
Methods changed in the Modifier class: modify_flows.
Methods changed in the Detector class: execute_classifiers.

All methods were modified to execute the model without some steps
that were performed in the training model. For this purpose, a new
boolean parameter was added to each of them.

Two new methods were added in the Gatherer class. The
check_path_exist method to create the required directories and
nfcapd_collector to collect all data exported by OpenFlow switches.

The get_ma_models method was excluded because the instance of the
class Detector is created at the beginning of the program allowing to
maintain the properties of the model without the need to return them
to the controller.