Modify the aggregation_flows and create_fetures methods
The aggregation_flows method was modified to establish a threshold when
the flows are aggregated.

The create_fetures method was removed from the modify_flows method in
order to create new features only after the aggregation method is
executed.

Several adjustments were made and the record_datatime and
processing_time methods were called in the controller and model
modules.