Add the menu and show_directory_content methods
A new module called tools was created in order to add some useful
functionality like the menu method. This method was implemented where
there was a while loop with options.

The show_directory_content method was part of the directory_content
method in the Gatherer class. This change was due to a refactoring in
the directory_content method.

The average parameter of the precision, recall and f1 score metrics
was removed. This parameter was causing an incorrect evaluation of
machine learning algorithms.