Remove some parameters from the classifiers
The choice to remove some parameters was due to the processing time
of the classifiers. For example, in this type of dataset, the linear
kernel of the support vector machine takes a long time to process the
training data, and in the end, the rbf kernel is always chosen by the
grid search method.

The cross-validator used in the grid search was changed to stratified
k-folds, the default parameter.