Add the creation module
This module defines the creation blueprint and three types of views.

The content view is responsible for showing all directories and files. 
Directories can be opened to further explore the content within them. 
When a file is selected, the next view is called to display the 
appropriate parameters.

The parameter view is responsible for displaying the parameters 
according to the function passed as arguments in the url. When the 
parameters are defined, a specific function view is called.

Each of the function views performs a step in the process of creating 
the dataset to be used in the anomaly detection.