Refactor results route
All changes made to the results route were made due to the 
implementation of the ORM model and retraining. Model persistence was 
added in another route called model.

The model route is similar to the result routes in some steps, with the 
main difference being the model retrain for persistence.