Refactor preprocess module
The staticmethod decorator was removed from the Formatter class so that 
only the format_header and format_flows methods are used. New boolean 
attributes were added to signal when specific processes are occurring.

All changes made to the Modifier class were to make it possible to 
aggregate one flow at a time instead of all. This procedure is essential 
in RealtimeThread when intrusions should be blocked. Waiting for all 
flows to aggregate can take a long time, allowing the attack to 
continue.

The Extractor class was changed to extract one flow at a time similar to 
the Modifier class procedure.