Refactor setting views
The Looger object was added in almost every view to keep up with the 
current steps.

The preprocessing name was added to create the model file name.

The new wrapper methods of the Formatter and Modifier class were used.

The deep copy method was used to create a new classifier object to avoid 
errors when a user returns to the page.